Create Bank Term Deposit
The investment and portfolio department of the Bank of Portugal has been conducting direct marketing campaigns to identify potential customers who would subscribe to their term deposits. The objective of our analysis is to develop a robust machine-learning algorithm to predict if a potential client will subscribe to the Bank’s term deposits.